+++ src/app/auth.js
@@ -10,6 +10,12 @@ const initializePortal = async () => {
     //ici on initialise et retourne de type de compte portal connecte
     await portal.initialize();
     const context = portal.getContext();
+    const fontFamily = context.app.theme.typography.fontFamily;
+
+    // Appliquer la police d'écriture à l'ensemble de l'application
+    if (fontFamily) {
+      document.body.style.fontFamily = fontFamily;
+    }
     const accountType = context.app.extra.administrator.organization.resource;
     return accountType;
